Split tests into current-user vs users
